backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-modeltranslation>=0.19.16",
    "django-rest-framework>=0.1.0",
    "django-split-settings>=1.3.2",
    "django-unfold>=0.63.0",
    "djangorestframework-api-key==2.*",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.15.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.0",
]
